+++ src/image/libvips.hpp
@@ -35,7 +35,7 @@ public:
 
     [[nodiscard]] auto width() const -> int override;
     [[nodiscard]] auto height() const -> int override;
-    [[nodiscard]] auto size() const -> uint64_t override;
+    [[nodiscard]] auto size() const -> size_t override;
     [[nodiscard]] auto data() const -> const unsigned char* override;
     [[nodiscard]] auto channels() const -> int override;
 
@@ -59,7 +59,7 @@ private:
 
     int max_width;
     int max_height;
-    uint64_t _size;
+    size_t _size;
 
     // for animated pictures
     int top = 0;
